requirements.txt
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.10
anaconda-client==1.6.14
anaconda-navigator==1.8.7
anaconda-project==0.8.2
appdirs==1.4.4
argcomplete==1.10.3
asn1crypto==0.24.0
astor==0.8.1
astroid==1.6.3
astropy==3.0.2
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==18.1.0
Babel==2.5.3
backcall==0.1.0
backports.entry-points-selectable==1.1.1
backports.shutil-get-terminal-size==1.0.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
blis==0.7.5
bokeh==0.12.16
boto==2.48.0
Bottleneck==1.2.1
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.4
catalogue==2.0.6
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
charset-normalizer==2.0.9
clang==5.0
click==7.1.2
cloudpickle==2.0.0
clyent==1.2.2
colorama==0.3.9
compressed-rtf==1.0.6
comtypes==1.1.4
conda==4.5.4
conda-build==3.10.5
conda-verify==2.0.0
conllu==4.4
contextlib2==0.5.5
contextvars==2.4
cryptography==2.2.2
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.6
Cython==0.28.2
cytoolz==0.9.0.1
dask==0.17.5
dataclasses==0.8
datashape==0.5.4
decorator==4.3.0
distlib==0.3.4
distributed==1.21.8
django-block-ip==0.2.0
dlib==19.22.1
dm-tree==0.1.6
docutils==0.14
docx2txt==0.8
ebcdic==1.1.1
EbookLib==0.17.1
efficientnet==1.1.1
entrypoints==0.2.3
et-xmlfile==1.0.1
extract-msg==0.28.7
fake-useragent==0.1.11
fastcache==1.0.2
feedparser==6.0.8
filelock==3.4.0
Flask==1.0.2
Flask-Cors==3.0.4
flatbuffers==1.12
frozenlist==1.2.0
gast==0.4.0
gevent==1.3.0
glob2==0.6
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==0.4.13
grpcio==1.42.0
gTTS==1.2.0
gTTS-token==1.1.4
h5py==2.10.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.6
idna-ssl==1.1.0
image-classifiers==1.0.0
imageio==2.3.0
imagesize==1.0.0
IMAPClient==2.1.0
imbalanced-learn==0.8.1
imblearn==0.0
immutables==0.16
importlib-metadata==4.8.2
importlib-resources==5.4.0
imutils==0.5.4
ipcalc==1.99.0
ipykernel==4.8.2
ipython==6.4.0
ipython_genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
joblib==1.1.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libretranslatepy==2.1.1
libs==0.0.10
llvmlite==0.23.1
locket==0.2.0
lxml==4.2.1
Markdown==3.3.6
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
mock==4.0.3
more-itertools==4.1.0
mpmath==1.0.0
msgpack==1.0.3
msgpack-numpy==0.4.4
msgpack-python==0.5.6
multidict==5.2.0
multipledispatch==0.5.0
murmurhash==1.0.6
mysql-connector-python==8.0.27
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.5.1
nltk==3.3
nmslib==2.1.1
nose==1.3.7
notebook==5.5.0
numba==0.38.0
numexpr==2.6.5
numpy==1.19.5
numpydoc==0.8.0
oauthlib==3.1.1
odo==0.5.1
olefile==0.46
opencv-python==4.5.4.60
openpyxl==2.5.3
opt-einsum==3.3.0
packaging==21.3
pandas==0.23.0
pandocfilters==1.4.2
parse==1.19.0
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
pathy==0.6.0
patsy==0.5.0
pdfminer.six==20191110
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.1.0
pkginfo==1.4.2
plac==0.9.6
platformdirs==2.4.0
playsound==1.2.2
pluggy==0.6.0
ply==3.11
preshed==3.0.6
prompt-toolkit==1.0.15
protobuf==3.19.1
psutil==5.4.5
py==1.5.3
py4j==0.10.9.2
pyAesCrypt==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.1
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.12.0
pycurl==7.43.0.1
pydantic==1.7.4
pyee==8.2.2
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.4
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
PyPDF2==1.26.0
pyppeteer==0.2.6
pyquery==1.4.3
pysbd==0.3.4
PySocks==1.6.8
pyspark==3.2.0
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-dateutil==2.7.3
python-pptx==0.6.21
pytz==2018.4
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.1.1
pywin32==223
pywinpty==0.5.1
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
regex==2018.1.10
requests==2.26.0
requests-html==0.10.0
requests-oauthlib==1.3.0
rope==0.10.7
rsa==4.8
ruamel_yaml==0.15.35
scikit-image==0.17.2
scikit-learn==0.24.2
scikit-surprise==1.1.1
scipy==1.1.0
scispacy==0.4.0
seaborn==0.8.1
segmentation-models @ git+https://github.com/qubvel/segmentation_models@94f624b7029deb463c859efbd92fa26f512b52b8
Send2Trash==1.5.0
sgmllib3k==1.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
slack==0.0.2
slackclient==2.9.3
smart-open==5.2.1
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
soupsieve==2.3.1
spacy==3.0.1
spacy-legacy==3.0.8
SpeechRecognition==3.8.1
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.2.7
srsly==2.4.2
statsmodels==0.9.0
surprise==0.1
sympy==1.1.1
tables==3.4.3
tblib==1.3.2
tensorboard==1.13.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-probability==0.15.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
textract==1.6.4
thinc==8.0.13
threadpoolctl==3.0.0
tifffile==2020.9.3
toolz==0.9.0
torch==1.10.2
tornado==5.0.2
tqdm==4.62.3
traitlets==4.3.2
translate==3.6.1
typer==0.3.2
typing==3.6.4
typing-extensions==3.7.4.3
tzdata==2021.5
tzlocal==4.1
unicodecsv==0.14.1
urllib3==1.26.8
virtualenv==20.10.0
virtualenvwrapper-win==1.2.6
w3lib==1.22.0
wasabi==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==9.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
win-unicode-console==0.5
win_inet_pton==1.0.1
wincertstore==0.2
wit==6.0.0
wrapt==1.12.1
xgboost==1.5.2
xlrd==1.2.0
XlsxWriter==1.0.4
xlwings==0.11.8
xlwt==1.3.0
yahoo-fin==0.8.9.1
yarl==1.7.2
zict==0.1.3
zipp==3.6.0